refactor(api): extract run execution from the HTTP path (stages #50)
Splits the chat loop along the SPEC §9.5 seam:

- RunExecutionService (new, transport-agnostic): context assembly at
  execution time (compaction summary + live window up to the triggering
  message), the model call, and every durable side effect — assistant
  turn persistence, run lifecycle + coalesced model.delta events, and
  post-turn compaction/titling. No HTTP types; the caller supplies the
  ModelClient, so the API keeps its 402-before-persistence contract and
  the worker will resolve credentials itself.
- ChatLoopService keeps only the API-side steps: validate, store the
  user message, create the run (same transaction), hand off execution.

Today the hand-off returns the live stream to the controller; the
worker move (#50) replaces that one call with an enqueue while the
worker drives the identical executeRun. The only observable change is
that context assembly now reads in its own transaction after the
message commit (read-committed makes this equivalent).

Behavior-preserving by proof: the full e2e suite (event ordering,
titling, abort/retry, replay, cross-tenant denial) and RLS integration
pass unchanged, plus lint/build/unit suites.

Refs #48, #50
