+++ CHANGELOG.md
@@ -2,6 +2,7 @@ _Reverse-chronological record of shipped work — features, fixes, and chores. N
 
 # 2026-07-02
 
+- Extracted run execution out of the HTTP path (staging #50): a new transport-agnostic `RunExecutionService` owns context assembly, the model call, and every durable side effect (assistant turn, run lifecycle + delta events, post-turn compaction/titling); `ChatLoopService` shrinks to the SPEC §9.5 API-side steps — validate, store message, create run, hand off. Behavior-preserving (full e2e parity); the worker move (#50) now swaps one hand-off call for an enqueue.
 - Added test CI (#70): a GitHub Actions workflow gates every PR (and pushes to `master`) on `turbo run lint`, `turbo run build`, the api unit suite, and `apps/api/scripts/rls-test.sh` — the cross-tenant RLS proof and HTTP e2e against a throwaway Postgres, same script as local. Actions are SHA-pinned, `permissions: contents: read`, actionlint + zizmor clean. Standing up root lint surfaced that `packages/ui`'s lint had been silently broken forever (no `eslint` devDependency) — fixed, along with the three warnings it had been hiding.
 - Made durable runs observable and replayable (#48/#49 API side): the loop now persists coalesced `model.delta` events (size-buffered via a pure delta-buffer, ordered by a sequential write chain), and a new run read surface exposes `GET /api/v1/runs/:id` plus the SPEC §9.4 cursor SSE `GET /api/v1/runs/:id/events?after_sequence=N` — each frame's SSE `id:` is its event sequence, an in-flight run is polled until terminal, a finished run streams its tail and closes, and a reconnect resumes from the last id with nothing lost. Cross-tenant reads 404 on both endpoints (proven in e2e). The `apps/web` resume-on-refresh client remains open in #49.
 - Landed the durable-run substrate (#48, first slice): `runs` and append-only `run_events` tables (SPEC §9.3–§9.4) with RLS `ENABLE`+`FORCE` and cross-tenant read/write denial proven live. Every user message now creates a run **in the same transaction** as the message, and the streaming loop dual-writes an ordered lifecycle log (`run.created` → `run.started` → `model.requested` → `model.completed` → `run.completed`/`run.failed`/`run.cancelled`) — the durable source of truth the SSE replay (#49) will read. Still to come in #48: the worker consuming from pg-boss, token-delta events, cancellation/heartbeat/timeout, and per-chat single-flight (deliberately deferred until heartbeat exists, since without it a crashed run would deadlock its chat).
