+++ apps/api/src/chats/run-execution.service.ts
@@ -1,0 +1,354 @@
+import { Injectable, Logger } from '@nestjs/common';
+import type { ModelMessage as AiModelMessage } from 'ai';
+
+import { TenantDbService } from '../db/tenant-db.service';
+import { type Message } from '../db/schema';
+import { type ModelClient } from '../models/model-client';
+import { CompactionsRepository, MessagesRepository } from './chats-repository';
+import { CompactionService } from './compaction.service';
+import {
+  buildContext,
+  DEFAULT_MAX_MESSAGES,
+  partsToText,
+  type MessagePart,
+  type StoredMessage,
+} from './context-builder';
+import { createDeltaBuffer } from './delta-buffer';
+import { RunEventsRepository, RunsRepository } from './runs-repository';
+import { TitleService } from './title.service';
+import {
+  buildTurnTelemetry,
+  emitCompletedTurnTelemetryLog,
+  turnTelemetryLogger,
+  type TurnTelemetry,
+} from './turn-telemetry';
+
+export const CHAT_SYSTEM_PROMPT =
+  'You are llame, an answer-only assistant. Answer the latest user message directly. Do not claim to use tools or take external actions.';
+
+/** The already-persisted user turn a run executes against. */
+export type RunUserMessage = {
+  id: string;
+  seq: number;
+  parts: MessagePart[];
+};
+
+/**
+ * RunExecutionService (#48/#50, SPEC §9.5) — executes one run: context
+ * assembly, the model call, and every durable side effect (assistant turn,
+ * run lifecycle + model.delta events, post-turn compaction/titling).
+ *
+ * Extracted from the HTTP-coupled ChatLoopService so the exact same execution
+ * drives both venues: today the request thread returns the live stream to the
+ * controller; the queue worker (#50) will call executeRun and consume the
+ * stream itself. Everything here is transport-agnostic — no HTTP types, and
+ * the caller supplies the ModelClient (credential resolution stays with the
+ * caller: the API's 402-before-persistence contract, the worker's own resolve).
+ */
+@Injectable()
+export class RunExecutionService {
+  private readonly logger = new Logger(RunExecutionService.name);
+
+  constructor(
+    private readonly tenantDb: TenantDbService,
+    private readonly compaction: CompactionService,
+    private readonly titles: TitleService,
+  ) {}
+
+  async executeRun(input: {
+    runId: string;
+    chatId: string;
+    userId: string;
+    userMessage: RunUserMessage;
+    client: ModelClient;
+    abortSignal?: AbortSignal;
+  }): Promise<ReturnType<ModelClient['streamText']>> {
+    const { client } = input;
+
+    // Context assembly happens at execution time (not enqueue time): the run
+    // reads the chat as it exists when it starts — compaction summary + live
+    // window up to the triggering message (SPEC §9.5 puts this worker-side).
+    const { system, messages } = await this.tenantDb.runAs(
+      input.userId,
+      async (tx) => {
+        const compaction = await new CompactionsRepository(
+          tx,
+        ).findLatestByChatId(input.chatId, input.userId);
+
+        const history = await new MessagesRepository(tx).findByChatId(
+          input.chatId,
+          input.userId,
+          {
+            maxSeq: input.userMessage.seq,
+            ...(compaction ? { sinceSeq: compaction.uptoSeq } : {}),
+            limit: DEFAULT_MAX_MESSAGES,
+          },
+        );
+
+        return buildContext(history as StoredMessage[], {
+          systemPrompt: CHAT_SYSTEM_PROMPT,
+          maxMessages: DEFAULT_MAX_MESSAGES,
+          ...(compaction
+            ? {
+                compaction: {
+                  summary: compaction.summary,
+                  uptoSeq: compaction.uptoSeq,
+                },
+              }
+            : {}),
+        });
+      },
+    );
+
+    const streamStartedAt = Date.now();
+
+    // Durable run lifecycle (#48, SPEC §9.4): the run row + run.created were
+    // written with the user message; execution events follow here. While the
+    // loop still executes on the request thread (until #50), these writes are
+    // observability dual-writes — they must never break the live stream.
+    await this.recordRunProgress(input.userId, input.runId, async (tx) => {
+      await new RunsRepository(tx).markStarted(input.runId, input.userId);
+      const events = new RunEventsRepository(tx);
+      await events.append(input.runId, 'run.started');
+      await events.append(input.runId, 'model.requested', {
+        model: client.model,
+        provider: client.provider,
+      });
+    });
+
+    // model.delta persistence (#48/#49): deltas are coalesced (delta-buffer)
+    // and appended through a sequential promise chain so events land in stream
+    // order even though onTextDelta fires synchronously.
+    const deltas = createDeltaBuffer();
+    let deltaWrites: Promise<void> = Promise.resolve();
+    const persistDelta = (text: string | null) => {
+      if (text === null) {
+        return;
+      }
+      deltaWrites = deltaWrites.then(() =>
+        this.recordRunProgress(input.userId, input.runId, async (tx) => {
+          await new RunEventsRepository(tx).append(input.runId, 'model.delta', {
+            text,
+          });
+        }),
+      );
+    };
+
+    return client.streamText({
+      system,
+      messages: messages as AiModelMessage[],
+      abortSignal: input.abortSignal,
+      onTextDelta: (text) => persistDelta(deltas.push(text)),
+      onError: async ({ error }) => {
+        // On the request thread the stream has already sent HTTP headers, so
+        // this error can't reach an exception filter — log + record it.
+        this.logger.error(
+          `Stream error for chat ${input.chatId}`,
+          error instanceof Error ? error.stack : String(error),
+        );
+
+        const telemetry = buildTurnTelemetry({
+          usage: null,
+          finishReason: null,
+          status: input.abortSignal?.aborted ? 'aborted' : 'error',
+          model: client.model,
+          provider: client.provider,
+          latencyMs: Date.now() - streamStartedAt,
+        });
+
+        await this.recordAssistantTurn({
+          chatId: input.chatId,
+          userId: input.userId,
+          inReplyTo: input.userMessage.id,
+          parts: [],
+          telemetry,
+        });
+
+        const status = telemetry.status === 'aborted' ? 'cancelled' : 'failed';
+        persistDelta(deltas.flush());
+        await deltaWrites;
+        await this.recordRunProgress(input.userId, input.runId, async (tx) => {
+          await new RunEventsRepository(tx).append(input.runId, 'run.failed', {
+            status,
+            message: error instanceof Error ? error.message : String(error),
+          });
+          await new RunsRepository(tx).markFinished(
+            input.runId,
+            input.userId,
+            status,
+            {
+              message: error instanceof Error ? error.message : String(error),
+            },
+          );
+        });
+      },
+      onFinish: async ({ text, usage, finishReason }) => {
+        const telemetry = buildTurnTelemetry({
+          usage,
+          finishReason,
+          status: input.abortSignal?.aborted
+            ? 'aborted'
+            : finishReason === 'error'
+              ? 'error'
+              : 'completed',
+          model: client.model,
+          provider: client.provider,
+          latencyMs: Date.now() - streamStartedAt,
+        });
+
+        await this.recordAssistantTurn({
+          chatId: input.chatId,
+          userId: input.userId,
+          inReplyTo: input.userMessage.id,
+          parts: [{ type: 'text', text }],
+          telemetry,
+        });
+
+        const status =
+          telemetry.status === 'completed' ? 'completed' : 'cancelled';
+        // Drain buffered deltas BEFORE the terminal events so the log reads
+        // in stream order: …model.delta, model.completed, run.completed.
+        persistDelta(deltas.flush());
+        await deltaWrites;
+        await this.recordRunProgress(input.userId, input.runId, async (tx) => {
+          const events = new RunEventsRepository(tx);
+          await events.append(input.runId, 'model.completed', {
+            usage,
+            finishReason,
+          });
+          await events.append(input.runId, `run.${status}`);
+          await new RunsRepository(tx).markFinished(
+            input.runId,
+            input.userId,
+            status,
+          );
+        });
+
+        // Post-turn work (#57 compaction, #78 titling): fire-and-forget — never
+        // delays or fails the finished turn. Both ride with the loop into the
+        // worker (#50).
+        if (telemetry.status === 'completed') {
+          void this.compaction.maybeCompact({
+            chatId: input.chatId,
+            userId: input.userId,
+            client,
+          });
+          void this.titles.maybeGenerateTitle({
+            chatId: input.chatId,
+            userId: input.userId,
+            client,
+            userText: partsToText(input.userMessage.parts),
+          });
+        }
+      },
+    });
+  }
+
+  /**
+   * Best-effort run bookkeeping (#48). While the loop runs on the request
+   * thread, run rows/events are a durability dual-write: failures are logged,
+   * never surfaced into the live stream. When the loop moves into the worker
+   * (#50) these become the authoritative execution record.
+   */
+  private async recordRunProgress(
+    userId: string,
+    runId: string,
+    write: (
+      tx: Parameters<Parameters<TenantDbService['runAs']>[1]>[0],
+    ) => Promise<void>,
+  ): Promise<void> {
+    try {
+      await this.tenantDb.runAs(userId, write);
+    } catch (error) {
+      this.logger.error(
+        `Failed to record run progress for run ${runId}`,
+        error instanceof Error ? error.stack : String(error),
+      );
+    }
+  }
+
+  private async recordAssistantTurn(input: {
+    chatId: string;
+    userId: string;
+    inReplyTo: string;
+    parts: MessagePart[];
+    telemetry: TurnTelemetry;
+  }): Promise<void> {
+    try {
+      const assistantMessage = await this.persistAssistantMessage(input);
+
+      if (assistantMessage) {
+        emitCompletedTurnTelemetryLog(turnTelemetryLogger, {
+          chatId: input.chatId,
+          messageId: assistantMessage.id,
+          inReplyTo: input.inReplyTo,
+          telemetry: input.telemetry,
+          onError: (error) => {
+            this.logger.error(
+              `Failed to emit assistant turn telemetry for chat ${input.chatId}`,
+              error instanceof Error ? error.stack : String(error),
+            );
+          },
+        });
+      }
+    } catch (error) {
+      this.logger.error(
+        `Failed to persist assistant turn for chat ${input.chatId}`,
+        error instanceof Error ? error.stack : String(error),
+      );
+    }
+  }
+
+  private async persistAssistantMessage(input: {
+    chatId: string;
+    userId: string;
+    inReplyTo: string;
+    parts: MessagePart[];
+    telemetry: TurnTelemetry;
+  }): Promise<Message | undefined> {
+    return this.tenantDb.runAs(input.userId, async (tx) => {
+      const messagesRepo = new MessagesRepository(tx);
+      const turn = await messagesRepo.findTurnState(
+        input.chatId,
+        input.userId,
+        input.inReplyTo,
+      );
+
+      if (turn.assistantMessage) {
+        if (isCompletedAssistantTurn(turn.assistantMessage)) {
+          return undefined;
+        }
+
+        return messagesRepo.updateAssistantReply({
+          id: turn.assistantMessage.id,
+          chatId: input.chatId,
+          inReplyTo: input.inReplyTo,
+          parts: input.parts,
+          usage: input.telemetry,
+        });
+      }
+
+      // The user turn must still exist (it was persisted before streaming). If it's gone
+      // — e.g. the chat was deleted mid-stream — skip rather than hit an in_reply_to FK error.
+      if (!turn.userMessage) {
+        return undefined;
+      }
+
+      return messagesRepo.createAssistantReplyIfAbsent({
+        chatId: input.chatId,
+        parts: input.parts,
+        usage: input.telemetry,
+        inReplyTo: input.inReplyTo,
+      });
+    });
+  }
+}
+
+export function isCompletedAssistantTurn(message: Message): boolean {
+  const usage = message.usage;
+  if (typeof usage !== 'object' || usage === null || !('status' in usage)) {
+    return true;
+  }
+
+  return (usage as { status?: unknown }).status === 'completed';
+}
